Day04 complete
I meant to do this earlier but better late than never.
My earlier puzzles are very brute force and could be cleaned up quite a bit, as some of today as well as it isn't optimized. At this point I'm going for speed over elegance with the plan to refactor these later.

README is a symlink to the index notebook we use as our Table of Contents. I don't quite know if this will work but if it does, it would be a great way to just synchronize Github's parsing with the convenience of using Livebook to navigate to each puzzle.

I also wanted to highlight the AOC helper and the somewhat hidden template. Creating the badge makes it far easier to work with.