Improved Index/Readme
I realized I should create a new project for each year. That may pad my repository count but I already ignore most of them. It really keeps the complexity contained and allows me to potentially redo all of this next year.
I also make it a point to go for elegant code, not speed. My brute force attempts for days 1-3 should be refactored at some point. This is more about learning for me.
I typically don't refactor something until I do it at least twice, that helps me see the abstraction boundaries.

Puzzles stay at the top above the Table of Contents that includes it. That may be confusing but I wanted the ToC to cover the setup and usage.
Add an installation section to set up the environment.
Add a usage section to isolate using the Livebooks and creating new ones. Future me would likely just use this project, not attempt to go through the workflow portion of solving the puzzles.
Add a workflow section to expand on how I work through each day.

I mentioned adding a separate PartTwo introduction because I'm not sure how the template approaches that. I did like keeping the description markdown closer to the code but I haven't taken the other approach yet.

I'm also realizing I don't know if this is standardizing on the right word choices or should workflow be solve the puzzles? Naming is hard y'all.